+++ src/lib.rs
@@ -3,7 +3,7 @@
 use anyhow::Context;
 use secp256k1::hashes::Hash;
 
-pub mod address_verifier;
+pub mod compliance;
 pub mod capped_hashmap;
 pub mod committee;
 pub mod constants;
